// src/lib/api/models/CreateSkuRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateSkuRequest = {
    order_product_code: string;
    invoice_product_code: string;
    description: string;
    package: string;
    commission_percentage: number;
};