// src/lib/api/models/UpdateSkuRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateSkuRequest = {
    order_product_code?: (string | null);
    invoice_product_code?: (string | null);
    description?: (string | null);
    package?: (string | null);
    commission_percentage?: (number | null);
};